requirements.txt
numpy
pandas
sklearn
flask
gunicorn
PyYAML
evidently